修正jdk1.6兼容性问题

+++ fc/org.tinygroup.flowbasiccomponent/src/test/java/org/tinygroup/flowbasiccomponent/test/testcase/FlowTest.java
@@ -27,7 +27,7 @@ public class FlowTest extends TestCase{
 		c.put("i", 1);
 		c.put("j", 2);
 		FlowTestUtil.executePage("callComputePlus", c);
-		int result=c.get("plusResult");
+		int result=(Integer) c.get("plusResult");
 		assertEquals( 3,result);
 	}
 	public void testCallBeanMethod(){
@@ -40,7 +40,7 @@ public class FlowTest extends TestCase{
 		Context c = new ContextImpl();
 		c.put("age", 12);
 		FlowTestUtil.executePage("callEl", c);
-		int result=c.get("elResult");
+		int result=(Integer) c.get("elResult");
 		assertEquals( 13,result);
 	}
 	public void testContextToObject(){
