+++ framework/org.tinygroup.service/src/test/java/org/tinygroup/service/test/testcase/UserServiceTest.java
@@ -72,7 +72,7 @@ public class UserServiceTest extends TestCase {
 		context.put("name", "testuser");
 		context.put("age", "1");
 		ServiceTestUtil.execute("serviceSetServiceUserAge", context);
-		int i=context.get("i");
+		int i=(Integer) context.get("i");
 		assertEquals(1,i);
 	}
 
@@ -95,7 +95,7 @@ public class UserServiceTest extends TestCase {
 		context.put("names", new String[5]);
 		context.put("ages", new int[5]);
 		ServiceTestUtil.execute("serviceSetServiceUserAgeArray", context);
-		int length=context.get("length");
+		int length=(Integer)context.get("length");
 		assertEquals(5,length);
 	}
